fix: improve source file format detection (#282)
The source file format detection was not working for e.g. `.tgz` files,
because their MIME type was not detected correctly.
I added the right MIME type definition for the `.tgz` files, but then I
decided to fix one old TODO and rework the file format detection a bit.

Now if we can't detect the MIME type of the file, we don't automatically
assume it to be text, but instead we try to detect its encoding from the
contents, using the `isTextOrBinary` package.
It should be relatively fast, since:
- it's done only for files which couldn't be determined via MIME type
- it's done only when the total size of the pushed files is <3MB
- it reads only a few bytes in the start, middle and end of each file,
not the whole file, and checks if they are in the Unicode characters
range, or not

This should not change the behavior for files which were really text, it
should just detect binary files better, so it's not a breaking change.