feat: Add hidden `edit-input-schema` command (#286)
Added a hidden command `apify edit-input-schema <PATH>`, which will open
the input schema editor at
https://apify.github.io/input-schema-editor-react, send the input schema
from PATH to it, and start an express server and listen to the schema
updates from the editor, and save them to the disk.

It is a hidden command for now, so that we can test it internally and
decide if we want to release it to the public, or not. If we do, we
should fix the biggest bugs with the input schema editor, and improve
the docs for this command and write some basic docs for it.

We were also talking about building the editor into the Apify Console,
so if we do that, maybe this will need to be reworked a bit.