Fix updating actor.json from template in create command (#304)
I've noticed that there is a bug in `create` command.

When you call `~/apify create fooo` and pick template with
`.actor/actor.json` then the `getLocalConfig` always returns
`undefined`. It's because your working directory is not in actor dir -
so it tries to read `~/.actor/actor.json` instead of
`~/fooo/.actor/actor.json`.

As a result we always rewrite `actor.json` from template with
`EMPTY_LOCAL_CONFIG`.