feat: add a troubleshooting section to the root help screen (#491)
Recently, we deployed a bugged version where apify login would fail.

Context on Slack: 
https://apifier.slack.com/archives/C0L33UM7Z/p1708169059446719.

We caught this internally but should this be first experienced by an
external user, we want to provide them with some basic instructions on
what to do.

This adds a new `TROUBLESHOOTING` section to the root help screen with
the same content as was added in #484.

Note: The CLI was crashing when printing help with the latest oclif
packages. So the original oclif packages were kept and a slightly older
version of @oclif/core was installed.