chore: Update local dev script to allow running from anywhere (#513)
I was trying to run the local copy of the CLI from a different folder,
like `../apify-cli/bin/dev.js`, but unfortunately it didn't work,
because the shebang in the `dev.js` file was resolved relative to the
current working directory, instead of relative to the `apify-cli`
folder.

This adds a `dev.sh` script, which will work when called from anywhere,
and run `apify-cli/node_modules/.bin/tsx`. This allows you to test stuff
like `apify create`, `apify pull` etc. easier.