docs: pseudo-automated new reference markdown file (#747)
As requested by @TC-MO, I've pseudo-automated the generation of the
reference file that people see on our docs!

I will add in more documentation on how to alter this later (more
specifically handling new commands added to the CLI), as its a manual
mapping of commands (unfortunately couldn't do it fully automated due to
some commands having different positions across the categories)

There probably is a better way to handle this in the future, I just did
the speedy way for now. Ideally, we'd use oclif's systems to load all
commands, store some specific field on each command class and rely on
that to map commands around, but even then we'll need to handle custom
ordering which might make it janky. To be determined

---------

Co-authored-by: Michał Olender <92638966+TC-MO@users.noreply.github.com>