chore!: move from yargs to node:util (#871)
BREAKING CHANGE: The shorthand flag for `--input-file` in the `call` command has been changed from `-if` to `-f`.

Our argument parsing was changed from yargs to the built-in one from nodejs. This does come with some caveats (specifically aliases like `-if` will be interpreted as setting the flags `-i` and `-f` (if both are defined) or as `f` as the value of the `-i` flag.).