feat: properly implement the logic to use CLI with multiple backends
There were some commented stubs and leftover conditions in the code, but they didn't really
work properly. This would be really helpful for me for work on core, so when I found the stubs,
I decided to actually implement the support properly.

This required adding a new version of the auth.json file, I implemented the logic that it transparently
upgrades the old format to the new one.